Mochitests will fail if there is a failure
