+++ src/test/mochitest/browser_dbg-sourcemaps-reloading.js
@@ -53,4 +53,5 @@ add_task(async function() {
     }),
     "Breakpoint has correct line"
   );
+  ok(false, "fff")
 });
